backend/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
elevenlabs

# FastAPI Webhook Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
